Add a first version of the Japanese translation
I am not a native speaker so caution is advised. Ideally we would need
a native speaker to check the translation because I am sure there are
lots of mistakes and unnatural expressions.

Native Japanese speakers will hopefully be able to understand the meaning
the translation is supposed to convey and refine my sentence formulations
and vocabulary choices.